% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFix.R
\name{check_word}
\alias{check_word}
\title{Helper function that fixes words from user input}
\usage{
check_word(word)
}
\arguments{
\item{words}{}
}
\value{
list of correct words
}
\description{
Helper function that fixes words from user input
}
