% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFix.R
\name{fix_sentence}
\alias{fix_sentence}
\title{Helper function that fixes words from user input}
\usage{
fix_sentence(sentence)
}
\arguments{
\item{sentence}{}
}
\value{
list of correct words
}
\description{
Helper function that fixes words from user input
}
