% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFix.R
\name{fix_word}
\alias{fix_word}
\title{From word and letter can put the letter next to the target letter}
\usage{
fix_word(word, changed_word)
}
\arguments{
\item{word}{char}

\item{changed_word}{char}
}
\value{
new_word
}
\description{
From word and letter can put the letter next to the target letter
}
