% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{get_sentiment_score}
\alias{get_sentiment_score}
\title{function to calculate sentiment scores}
\usage{
get_sentiment_score(text)
}
\arguments{
\item{text}{vector}
}
\value{
scores
}
\description{
function to calculate sentiment scores
}
