% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFix.R
\name{getText}
\alias{getText}
\title{Gets text file from user and cleans it up to be ready to be used}
\usage{
getText()
}
\value{
clean_text
}
\description{
Gets text file from user and cleans it up to be ready to be used
}
