% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{iterated_difference}
\alias{iterated_difference}
\title{Helper Function, gets the iterated total difference}
\usage{
iterated_difference(data)
}
\arguments{
\item{data}{}
}
\value{
counts
}
\description{
Helper Function, gets the iterated total difference
}
