% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFix.R
\name{letter_find}
\alias{letter_find}
\title{Finds the difference in letters between two words inputted}
\usage{
letter_find(word, changed_word)
}
\arguments{
\item{word}{char}

\item{changed_word}{char}
}
\value{
difference a single letter
}
\description{
Finds the difference in letters between two words inputted
}
